+++ src/win.js
@@ -10,24 +10,51 @@ import { regexpPowerShell, shellRequiredError } from "./constants.js";
  * Escape a shell argument for use in CMD.
  *
  * @param {string} arg The argument to escape.
+ * @param {boolean} interpolation Is interpolation enabled.
  * @returns {string} The escaped argument.
  */
-function escapeShellArgsForCmd(arg) {
-  return arg.replace(/\u{0}/gu, "").replace(/"/g, `""`);
+function escapeShellArgsForCmd(arg, interpolation) {
+  let result = arg.replace(/\u{0}/gu, "");
+
+  if (interpolation) {
+    result = result
+      .replace(/\^/g, "^^")
+      .replace(/(<|>)/g, "^$1")
+      .replace(/(")/g, "^$1")
+      .replace(/(\&|\|)/g, "^$1");
+  } else {
+    result = result.replace(/"/g, `""`);
+  }
+
+  return result;
 }
 
 /**
  * Escape a shell argument for use in PowerShell.
  *
  * @param {string} arg The argument to escape.
+ * @param {boolean} interpolation Is interpolation enabled.
  * @returns {string} The escaped argument.
  */
-function escapeShellArgsForPowerShell(arg) {
-  return arg
+function escapeShellArgsForPowerShell(arg, interpolation) {
+  let result = arg
     .replace(/\u{0}/gu, "")
-    .replace(/("|“|”|„)/g, `$1$1`)
     .replace(/`/g, "``")
     .replace(/\$/g, "`$");
+
+  if (interpolation) {
+    result = result
+      .replace(/^((?:\*|[1-6])?)(>)/g, "$1`$2")
+      .replace(/^(<|@|#|-|\:|\])/g, "`$1")
+      .replace(/(,|\;|\&|\|)/g, "`$1")
+      .replace(/(\(|\)|\{|\})/g, "`$1")
+      .replace(/('|’|‘|‛|‚)/g, "`$1")
+      .replace(/("|“|”|„)/g, "`$1");
+  } else {
+    result = result.replace(/("|“|”|„)/g, "$1$1");
+  }
+
+  return result;
 }
 
 /**
@@ -35,15 +62,16 @@ function escapeShellArgsForPowerShell(arg) {
  *
  * @param {string} arg The argument to escape.
  * @param {string} shell The shell to escape the argument for.
+ * @param {boolean} interpolation Is interpolation enabled.
  * @returns {string} The escaped argument.
  */
-export function escapeShellArg(arg, shell) {
+export function escapeShellArg(arg, shell, interpolation) {
   if (shell === undefined) throw new TypeError(shellRequiredError);
 
   if (regexpPowerShell.test(shell)) {
-    return escapeShellArgsForPowerShell(arg);
+    return escapeShellArgsForPowerShell(arg, interpolation);
   } else {
-    return escapeShellArgsForCmd(arg);
+    return escapeShellArgsForCmd(arg, interpolation);
   }
 }