+++ test/index.fuzz.cjs
@@ -11,11 +11,13 @@ const shescape = require("../index.cjs");
 
 require("dotenv").config();
 
+const WHITESPACE_REGEX = /\s|\u0085/gu;
+
 function getFuzzShell() {
   return process.env.FUZZ_SHELL;
 }
 
-function prepareArg(arg) {
+function prepareArg(arg, quoted) {
   const shell = getFuzzShell();
   const isWindows = () => os.platform() === "win32";
   const isShellCmd = () => shell === undefined || /cmd\.exe$/.test(shell);
@@ -25,13 +27,22 @@ function prepareArg(arg) {
   if (isWindows()) {
     // Node on Windows ...
     if (isShellCmd()) {
-      // ... in CMD interprets arguments with `\"` as `"` (even if there's a
-      // null character between `\` and `"`) so we escape the `\`.
-      result = result.replace(/((\\\u{0}*)+)(?=\u{0}*("|$))/gu, "$1$1");
+      // ... in CMD, depending on if the argument is quotes ...
+      if (quoted) {
+        // ... interprets arguments with `\"` as `"` (even if there's a
+        // null character between `\` and `"`) so we escape the `\`.
+        result = result.replace(/((\\\u{0}*)+)(?=\u{0}*("|$))/gu, "$1$1");
+      } else {
+        // ... interprets arguments with `\"` as `"` so we escape the `\` ...
+        result = result.replace(/((\\\u{0}*)+)(?=\u{0}*")/gu, "$1$1");
+
+        // ... interprets arguments with `"` as `` so we escape it with `\`.
+        result = result.replace(/"/g, `\\"`);
+      }
     } else if (isShellPowerShell()) {
       // ... in PowerShell, depending on if there's whitespace in the
       // argument ...
-      if (/\s|\u0085/u.test(result)) {
+      if (WHITESPACE_REGEX.test(result)) {
         // ... interprets arguments with `""` as nothing so we escape it with
         // extra double quotes as `""""` ...
         result = result.replace(/"/g, `""`);
@@ -61,12 +72,28 @@ function getExpectedOutput(arg) {
     .replace(/\u{0}/gu, ""); // Remove null characters
 }
 
-function checkQuotesAndEscapesCorrectly(arg) {
-  const options = {
-    shell: getFuzzShell(),
-  };
+function checkEscapesCorrectly(arg, options) {
+  arg = arg.replace(WHITESPACE_REGEX, "");
+  const preparedArg = prepareArg(arg, false);
+  const escapedArg = shescape.escape(preparedArg, {
+    ...options,
+    interpolation: true,
+  });
+  const cmd = `node test/fuzz/echo.js ${escapedArg}`;
 
-  const preparedArg = prepareArg(arg);
+  const result = cp.execSync(cmd, options).toString();
+  const expected = getExpectedOutput(arg);
+  if (expected !== result) {
+    throw new Error(
+      "Unexpected output after escaping (- got, + expected):\n" +
+        `- ${result}\n` +
+        `+ ${expected}`
+    );
+  }
+}
+
+function checkQuotesAndEscapesCorrectly(arg, options) {
+  const preparedArg = prepareArg(arg, true);
   const quotedArg = shescape.quote(preparedArg, options);
   const cmd = `node test/fuzz/echo.js ${quotedArg}`;
 
@@ -74,7 +101,7 @@ function checkQuotesAndEscapesCorrectly(arg) {
   const expected = getExpectedOutput(arg);
   if (expected !== result) {
     throw new Error(
-      "Unexpected output (- got, + expected):\n" +
+      "Unexpected output after quoting and escaping (- got, + expected):\n" +
         `- ${result}\n` +
         `+ ${expected}`
     );
@@ -83,7 +110,12 @@ function checkQuotesAndEscapesCorrectly(arg) {
 
 function fuzz(buf) {
   const arg = buf.toString();
-  checkQuotesAndEscapesCorrectly(arg);
+  const options = {
+    shell: getFuzzShell(),
+  };
+
+  checkEscapesCorrectly(arg, options);
+  checkQuotesAndEscapesCorrectly(arg, options);
 }
 
 module.exports = {